CLN/DOC: refactored dummy testing methods into tests/dummy_wfmanager and documentation update
