+++ doc/source/api/force_wfmanager.model.tests.rst
@@ -13,18 +13,6 @@ force\_wfmanager\.model\.tests\.test\_analysis\_model module
     :show-inheritance:
 
 
-
-force\_wfmanager\.model\.tests\.test\_variable\_names\_registry module
---------------------------------------------------------------------------
-
-.. automodule:: force_wfmanager.model.tests.test_variable_names_registry
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-
-
 Module contents
 ---------------
