+++ doc/source/api/force_wfmanager.utils.rst
@@ -19,14 +19,6 @@ force\_wfmanager\.utils\.local\_traits module
     :undoc-members:
     :show-inheritance:
 
-force\_wfmanager\.utils\.task\_toggle\_group\_accelerator module
--------------------------------------------------------------------
-
-.. automodule:: force_wfmanager.utils.task_toggle_group_accelerator
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
 
 Module contents
 ---------------
